feat: add abstract stepper presenter
